Rename config package
