+++ apps/web/src/app/api/health/route.ts
@@ -5,7 +5,7 @@ import { headers } from "next/headers";
 
 import { rateLimiter } from "@/lib/rate-limit";
 import { ALLOWED_ORIGINS } from "@/lib/constants";
-import { HEALTH } from "@ffxiv-tools/config";
+import { HEALTH } from "../../../../../../packages/typescript-config/src";
 import { log } from "@/lib/logger";
 import { ServiceError } from "@/lib/errors";
