+++ apps/web/tsconfig.json
@@ -20,7 +20,7 @@
     ],
     "paths": {
       "@/*": ["./src/*"],
-      "@ffxiv-tools/config": ["../../packages/config/src"]
+      "@ffxiv-tools/config": ["../../packages/typescript-config/src"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "vitest.config.mts"],
