+++ package.json
@@ -14,7 +14,7 @@
     "build:web": "npm run build --workspace=@ffxiv-tools/web",
     "build:data-access": "cd services/data-access && cargo build --release",
     "build:storage": "cd services/storage && cargo build --release",
-    "build:config": "cd packages/config && tsc",
+    "build:config": "cd packages/typescript-config && tsc",
     "test": "concurrently \"npm run test:web\" \"npm run test:data-access\" \"npm run test:storage\"",
     "test:all": "npm run test:once --workspace=@ffxiv-tools/web && cd services/data-access && cargo test && cd ../storage && cargo test",
     "test:web": "npm run test --workspace=@ffxiv-tools/web",
