+++ packages/typescript-config/src/index.ts
